move TestCenterService from model/implementation to model, make it class service (singleton)
